pyproject.toml
[tool.poetry]
name = "screenshotscribe"
version = "0.1.0"
description = ""
authors = ["Derek Lee <22509961+yuandere@users.noreply.github.com>"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
screenshotscribe = "screenshotscribe.main:main"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.45.2"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"
python-docx = "^1.1.2"
tqdm = "^4.66.5"

